move intersections from jordan to piecewise
